Don't use octal numbers
When run with the "--use_strict" flag, NodeJS raises an error citing:

    SyntaxError: Octal literals are not allowed in strict mode.

when parsing files which contain Octal numbers (or numbers which look
like octals - usually because they have leading zeros). This commit
replaces two instances of this with strings.

See [this Stack Overflow answer](http://stackoverflow.com/a/23609255/574190) for more info.