Add a contentsByteOffset to the header emitted on entry events.
When emitting the header for an entry event, the contentsByteOffset
property is added to the header object. This property is the byte offset
of the start of the entry's file contents in the tar stream, after
the header.

This is useful when building an index of a tar's contents, for later
seeking to a specific entry's contents in the tar stream.

Headers can be variable length based on the tar implementation.